+++ CHANGES.md
@@ -3,6 +3,8 @@ Poodinis Changelog
 **Version NEXT**
 * ADD setting persistent registration options
 * DEPRECATE DO_NOT_ADD_CONCRETE_TYPE_REGISTRATION, use doNotAddConcreteTypeRegistration instead
+* ADD resolve options to container resolve()
+* ADD ability to register a type while resolving it. Use resolve option registerBeforeResolving
 
 **Version 6.0.0**
 * CHANGE registration scopes are replaced by a single factory implementation. If you were not doing anything with the internal scope mechanism, you